Web UI - Recording - After done, move to next sentence to prep for another recording. Fixes #48
